Moved Charge owner validation rule to correct directory
