+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Domain/Dtos/Validation/DocumentValidation/Factories/DocumentValidationRulesFactory.cs
@@ -14,7 +14,6 @@
 
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 using GreenEnergyHub.Charges.Domain.Dtos.Messages.Command;
 using GreenEnergyHub.Charges.Domain.Dtos.SharedDtos;
@@ -38,9 +37,7 @@ namespace GreenEnergyHub.Charges.Domain.Dtos.Validation.DocumentValidation.Facto
             var document = command.Document;
             ArgumentNullException.ThrowIfNull(document);
 
-            var chargeOwners = command.Operations.Select(o => o.ChargeOwner);
             var rules = await GetRulesForDocumentAsync(document).ConfigureAwait(false);
-
             return ValidationRuleSet.FromRules(rules);
         }
